+++ packages/workeros-programs/src/index.js
@@ -109,4 +109,5 @@ export const libraries = [
   { path: "/lib/workeros-node/net.js", source: () => fetchText("./node/net.js") },
   { path: "/lib/workeros-node/http.js", source: () => fetchText("./node/http.js") },
   { path: "/lib/workeros-node/crypto.js", source: () => fetchText("./node/crypto.js") },
+  { path: "/lib/workeros-node/zlib.js", source: () => fetchText("./node/zlib.js") },
 ];